move each data type to it's file
